fixed breadcrumbs bug and added notification, Icon
