+++ ui2/src/common/Modal.jsx
@@ -9,4 +9,3 @@ export const AppModal = ({ children, toggleModal, okDisabled, ...props }) => {
     </Modal>
   );
 };
-
