+++ ui2/src/utils/parse.js
@@ -1,0 +1,6 @@
+export const parseSearchUrl = searchUrl =>
+  searchUrl
+    .split('?')
+    .pop()
+    .split('&')
+    .map(searchParam => searchParam.split('=').shift());
